Work with TestRail was started

+++ src/main/java/com/solvd/dice/api/TestCase.java
@@ -4,11 +4,11 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public final class TestCase {
-    public final Integer id;
+    public final Long id;
     public final String title;
 
     @JsonCreator
-    public TestCase(@JsonProperty("id")Integer id, @JsonProperty("title")String title) {
+    public TestCase(@JsonProperty("id")Long id, @JsonProperty("title")String title) {
         this.id = id;
         this.title = title;
     }
