+++ src/main/java/com/solvd/dice/api/TestSuite.java
@@ -6,12 +6,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 public final class TestSuite {
-    public final Integer id;
+    public final Long id;
     public final String title;
     public final TestCase[] testCases;
 
     @JsonCreator
-    public TestSuite(@JsonProperty("id") Integer id,@JsonProperty("title") String title,@JsonProperty("childTestCases") TestCase[] testCases) {
+    public TestSuite(@JsonProperty("id") Long id,@JsonProperty("title") String title,@JsonProperty("childTestCases") TestCase[] testCases) {
         this.id = id;
         this.title = title;
         this.testCases = testCases;
